+++ Engine/Texture.h
@@ -27,8 +27,8 @@ public:
 
 	void Initiate(const TextureDesc& desc, const std::string& debug_name);
 	void Initiate(const TextureDesc& desc, bool create_from_texture, const std::string& debug_name);
-	void InitiateAsDepthStencil(float width, float height, const std::string& debug_name);
-	void InitiateAsRenderTarget(float width, float height, const std::string& debug_name);
+	void InitiateAsDepthStencil(s32 width, s32 height, const std::string& debug_name);
+	void InitiateAsRenderTarget(s32 width, s32 height, const std::string& debug_name);
 
 
 	ITexture2D* GetDepthTexture() const { return m_DepthTexture; }
